+++ quantify_scheduler/backends/graph_compilation.py
@@ -1,13 +1,18 @@
 # Repository: https://gitlab.com/quantify-os/quantify-scheduler
 # Licensed according to the LICENCE file on the main branch
 
-from typing import Callable, Dict, List, Optional, Union, Tuple
-import networkx as nx
+from typing import TYPE_CHECKING, Callable, Dict, List, Optional, Tuple, Union
+
 import matplotlib.pyplot as plt
+import networkx as nx
 from matplotlib.axes import Axes
-from quantify_scheduler.structure.model import DataStructure
-from quantify_scheduler import Schedule, CompiledSchedule
+
+from quantify_scheduler import CompiledSchedule, Schedule
 from quantify_scheduler.helpers.importers import import_python_object_from_string
+from quantify_scheduler.structure.model import DataStructure
+
+if TYPE_CHECKING:
+    from quantify_scheduler.device_under_test.quantum_device import QuantumDevice
 
 
 class CompilationError(RuntimeError):
@@ -184,7 +189,16 @@ class QuantifyCompiler(CompilationNode):
     compilation passes.
     """
 
-    def __init__(self, name):
+    def __init__(self, name, quantum_device: Optional["QuantumDevice"] = None) -> None:
+        """
+        Parameters
+        ----------
+        name
+            name of the compiler instance
+        quantum_device
+            quantum_device from which a :class:`~.CompilationConfig` will be generated
+            if None is provided for the compile step
+        """
         super().__init__(name=name)
 
         # current implementations use networkx directed graph to store the task graph
@@ -196,8 +210,10 @@ class QuantifyCompiler(CompilationNode):
         self._input_node = None
         self._ouput_node = None
 
+        self.quantum_device = quantum_device
+
     def compile(
-        self, schedule: Schedule, config: CompilationConfig
+        self, schedule: Schedule, config: Optional[CompilationConfig] = None
     ) -> CompiledSchedule:
         """
         Compile a :class:`~.Schedule` using the information provided in the config.
@@ -207,7 +223,8 @@ class QuantifyCompiler(CompilationNode):
         schedule
             the schedule to compile.
         config
-            describing the information required to compile the schedule.
+            describing the information required to compile the schedule. If not specified,
+            self.quantum_device will be used to generate the config.
 
         Returns
         -------
@@ -223,6 +240,10 @@ class QuantifyCompiler(CompilationNode):
 
         # classes inheriting from this node should overwrite the _compilation_func and
         # not the public facing compile.
+        if config is None:
+            if self.quantum_device is None:
+                raise RuntimeError("Either quantum_device or config must be specified")
+            config = self.quantum_device.generate_compilation_config()
         return self._compilation_func(schedule=schedule, config=config)
 
     @property